Rename route utils file to match default export
As the route utils file has the single responsibility for creating the route
actions, we can make it the default export and rename it to better communicate
it's use.
